fix(types): type cohersion on lhs of list repeat

+++ code/tst/python/gpu/contig.py
@@ -5,7 +5,7 @@ def fill(out: list[i32]) -> None:
     col = global_id(1)
     out[row * 5 + col] = 42
 
-out: list[i32] = [i32(0)] * 25
+out: list[i32] = [0] * 25
 
 fill(out, (5, 5, 1))
 for i in range(25):
